[BACKEND] fix comments.
