+++ .github/workflows/sunrise-build-and-test.yml
@@ -12,7 +12,7 @@ concurrency:
 
 jobs:
   sunrise-build-and-test:
-    runs-on: flagtree-sunrise
+    runs-on: sunrise3.6
     if: ${{ github.repository == 'FlagTree/flagtree' || github.repository == 'flagos-ai/flagtree' }}
     steps:
       - name: Setup environment
