+++ lib/Dialect/TritonGPU/IR/Dialect.cpp
@@ -2,7 +2,7 @@
 #include "flagtree_spec.h"
 #endif
 
-#ifndef FLAGTREE_SPEC_triton_Dialect_TritonGPU_IR_sunrise_Dialect
+#ifndef FLAGTREE_SPEC_Dialect_TritonGPU_IR_Dialect
 
 #include "triton/Dialect/Triton/IR/Dialect.h"
