+++ lib/Dialect/TritonGPU/IR/LinearLayoutConversions.cpp
@@ -2,7 +2,7 @@
 #include "flagtree_spec.h"
 #endif
 
-#ifndef FLAGTREE_SPEC_triton_Dialect_TritonGPU_IR_sunrise_LinearLayoutConversion
+#ifndef FLAGTREE_SPEC_Dialect_TritonGPU_IR_LinearLayoutConversion
 
 #include <vector>
